finish login/logout

+++ Back-end/src/tasks/tasks.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, InternalServerErrorException, Logger, NotFoundException } from '@nestjs/common';
 
 
 import { CreateTaskDto } from './dto/create-task.dto';
@@ -7,10 +7,12 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { Repository } from 'typeorm';
 import { TaskStatus } from './task-status.enum';
+import { User } from 'src/auth/user.entity';
 
 
 @Injectable()
 export class TasksService {
+    private logger = new Logger('TaskService');
     constructor(
        @InjectRepository(Task)
        private tasksRepository: Repository<Task>,
@@ -35,10 +37,15 @@ export class TasksService {
     //     return tasks;
     // }
 
-    async getTasks(filterDto: GetTasksFilterDto): Promise<Task[]> {
+    async getTasks(
+        filterDto: GetTasksFilterDto,
+        user: User,
+        ): Promise<Task[]> {
         const { status, search } = filterDto;
 
         const query = this.tasksRepository.createQueryBuilder('task');
+        
+        query.where('task.userId = :userId', { userId: user.id });
 
         if (status) {
             query.andWhere('task.status = :status', { status } );
@@ -48,12 +55,18 @@ export class TasksService {
             query.andWhere('(task.description LIKE :search OR task.title LIKE :search)', {search: `${search}`});
         }
 
-        const tasks = await query.getMany();
-        return tasks;
+        try{
+            const tasks = await query.getMany();
+            return tasks;
+        }catch(error) {
+            this.logger.error(`fail to get tasks for user "${user.username}". Filters: ${JSON.stringify(filterDto)}`, error.stack)
+            throw new InternalServerErrorException();
+        }
+
     }
 
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.tasksRepository.findOneBy({id: id});
+    async getTaskById(id: number, user: User): Promise<Task> {
+        const found = await this.tasksRepository.findOneBy({id: id, userId: user.id});
 
         if (!found) {
             throw new NotFoundException(`Task with ID "${id}" not found!`)
@@ -76,21 +89,35 @@ export class TasksService {
     //     return task;
     // }
 
-    async createTask(createTaskDto: CreateTaskDto) {
+    async createTask(
+        createTaskDto: CreateTaskDto,
+        user: User,
+        ): Promise<Task> {
         const { title, description } = createTaskDto;
 
         const task = new Task();
         task.title = title;
         task.description = description;
-        task.status = TaskStatus.OPEN
+        task.status = TaskStatus.OPEN;
+        task.userId = user.id;
+        task.user = user;
 
-        await this.tasksRepository.save(task);
-        
+        try{
+            await this.tasksRepository.save(task);
+        }catch(error){
+            this.logger.error(`Fail to create task for user ${user.username}. Task: ${JSON.stringify(createTaskDto)}`, error.stack)
+            throw new InternalServerErrorException();
+        }
+
+        delete task.user;
         return task;
     }
 
-    async deleteTask(id: number): Promise<void> {
-       const result = await this.tasksRepository.delete(id);
+    async deleteTask(
+        id: number,
+        user: User,
+        ): Promise<void> {
+       const result = await this.tasksRepository.delete({id, userId: user.id});
       
        if(result.affected === 0) {
         throw new  NotFoundException(`Task with ID "${id}" not found!`);
@@ -103,8 +130,12 @@ export class TasksService {
     //     return task;
     // }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(
+        id: number,
+        status: TaskStatus,
+        user: User,
+        ): Promise<Task> {
+        const task = await this.getTaskById(id, user);
         task.status = status;
         await this.tasksRepository.save(task);
         return task;
